Remove hard-coded output name and use file.path
The user can supply the name of the output folder as an argument, but inside the contrast_oper function, is assumes the output folder is "ballgown". If any other name is supplied, you'll run into 'directory does not exist' errors or results will be written to two different directories.

I also changed the use of paste(..., sep="\t") for path construction to the more ideomatic and platform-independent file.path(...).