Ensure that MGI library name parsing captures letter prefix.
That \w?_? in the regex is designed to catch the [ABCD] of the MGI sub-index, but is accidentally capturing the “S” from the barcode name when the sub-index is not present.

Given input like "ESO-20-1136TBIO-mid_treatment_2-1061787_SI_TT7_G11-SI_TT5_G11"

Before the fix, the match group <barcode_name> would capture "I_TT7_G11-SI_TT5_G11". Note the missing "S" at the beginning of the string, as it is accidentally captured by the \w?_?.

After the fix, the match group "barcode_name will capture "SI_TT7_G11-SI_TT5_G11", as we expect.