Fixes a bug caused by reading DGE results with NAs
In some cases, the adjusted p value calculated by deseq2 can be NA. This can occur when a feature/gene has a low mean normalised count. If the padj is NA, then the test for significance which involves summing a vector with NAs will return NA rather than a sum of values.

This small fix ensures that any features with padj==NA will be simply ignored when counting the number of significant features.